extract getPluginName and toCamelCase to lib-folder
